+++ Cargo.toml
@@ -33,6 +33,7 @@ headers = ["headers-shared", "headers-um", "headers-vc"]
     "shared-ntstatus",
     "shared-rpcndr",
     "shared-sspi",
+    "shared-usb",
     "shared-windef",
     "shared-windowsx",
     "shared-winerror",
@@ -50,6 +51,7 @@ headers = ["headers-shared", "headers-um", "headers-vc"]
 "shared-ntstatus" = ["shared-ntdef"]
 "shared-rpcndr" = []
 "shared-sspi" = ["shared-ntdef"]
+"shared-usb" = ["shared-minwindef", "um-winnt"]
 "shared-windef" = ["shared-minwindef", "um-winnt"]
 "shared-windowsx" = ["shared-minwindef"]
 "shared-winerror" = ["shared-minwindef"]
